@@ -1,8 +1,7 @@
 ﻿using BrandUp.Pages.Interfaces;
-using Newtonsoft.Json;
-using Newtonsoft.Json.Converters;
 using System.Collections.Generic;
 using System.ComponentModel.DataAnnotations;
+using System.Text.Json.Serialization;
 
 namespace BrandUp.Pages.Models
 {
@@ -20,7 +19,7 @@ namespace BrandUp.Pages.Models
         [Required]
         public string PageType { get; set; }
         [Required]
-        [JsonConverter(typeof(StringEnumConverter))]
+        [JsonConverter(typeof(JsonStringEnumConverter))]
         public PageSortMode Sort { get; set; }
     }
 }
