@@ -1,7 +1,6 @@
 ﻿using BrandUp.Pages.Interfaces;
-using Newtonsoft.Json;
-using Newtonsoft.Json.Converters;
 using System;
+using System.Text.Json.Serialization;
 
 namespace BrandUp.Pages.Models
 {
@@ -12,7 +11,7 @@ namespace BrandUp.Pages.Models
         public Guid? PageId { get; set; }
         public string Title { get; set; }
         public string PageType { get; set; }
-        [JsonConverter(typeof(StringEnumConverter))]
+        [JsonConverter(typeof(JsonStringEnumConverter))]
         public PageSortMode Sort { get; set; }
         public bool CustomSorting { get; set; }
         public string PageUrl { get; set; }
